CFIN-517 Refactor override URLs into constants files
