+++ src/constants/nursingOverrides.js
@@ -1,0 +1,12 @@
+const URLS_TO_OVERRIDE = [
+    "https://www.york.ac.uk/study/undergraduate/courses/bsc-nursing-adult/",
+    "https://www.york.ac.uk/study/undergraduate/courses/bsc-nursing-child/",
+    "https://www.york.ac.uk/study/undergraduate/courses/bsc-nursing-mental-health/",
+    "https://www.york.ac.uk/study/undergraduate/courses/mnurs-nursing-adult/",
+    "https://www.york.ac.uk/study/undergraduate/courses/mnurs-nursing-child/",
+    "https://www.york.ac.uk/study/undergraduate/courses/mnurs-nursing-mental-health/",
+];
+
+const OVERRIDE_URL = "https://www.york.ac.uk/study/undergraduate/subjects/nursing/";
+
+module.exports = {URLS_TO_OVERRIDE, OVERRIDE_URL};
